requirements.txt


Django==4.2.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
mysql-connector-python==8.1.0
PyMySQL==1.1.0
sqlparse==0.4.4
pipenv==2023.9.8
virtualenv==20.24.5
yarl==1.9.2


